+++ ChoGGi's Library/Changelog.md
@@ -4,6 +4,8 @@
 - Added IsArray.
 - ExportMapDataToCSV has option to return data in tables instead of csv.
 - CreateNumberEditor: Added param to skip creating XEdit.
+- Added: Examine/ExecCode/FindValue/ImageViewer/ObjectEditor (and Open* funcs).
+- Examine: Added tooltip_image/exec_tables.
 
 ## v9.6 (24 March 2021)
 - Added SetChoGGiPalette, IsGamepadButtonPressed.